refactor(Table): delete DisplayEntry
